fix due to creating config files in every place where user wants open application

+++ QuickViewFile/Helpers/ConfigHelper.cs
@@ -102,8 +102,11 @@ namespace QuickViewFile.Helpers
 
                 try
                 {
-                    string json = JsonSerializer.Serialize(configModel, new JsonSerializerOptions { WriteIndented = true });
-                    File.WriteAllText(ConfigFileName, json);
+                    if (File.Exists(ConfigFileName))
+                    {
+                        string json = JsonSerializer.Serialize(configModel, new JsonSerializerOptions { WriteIndented = true });
+                        File.WriteAllText(ConfigFileName, json);
+                    }
                 }
                 catch
                 {
